+++ Krbtgt/tests/pester.ps1
@@ -3,8 +3,9 @@
 	
 	$TestFunctions = $true,
 	
-	[ValidateSet('None', 'Default', 'Passed', 'Failed', 'Pending', 'Skipped', 'Inconclusive', 'Describe', 'Context', 'Summary', 'Header', 'Fails', 'All')]
-	$Show = "None",
+	[ValidateSet('None', 'Normal', 'Detailed', 'Diagnostic')]
+	[Alias('Show')]
+	$Output = "None",
 	
 	$Include = "*",
 	
@@ -16,13 +17,20 @@ Write-PSFMessage -Level Important -Message "Starting Tests"
 Write-PSFMessage -Level Important -Message "Importing Module"
 
 # Fake AD Module for automated testing in AzDev
-if ($env:SYSTEM_DEFAULTWORKINGDIRECTORY) {
+if ($env:SYSTEM_DEFAULTWORKINGDIRECTORY)
+{
 	New-Module -Name ActiveDirectory -ScriptBlock { } | Import-Module
 }
 
-Remove-Module Krbtgt -ErrorAction Ignore
-Import-Module "$PSScriptRoot\..\Krbtgt.psd1"
-Import-Module "$PSScriptRoot\..\Krbtgt.psm1" -Force
+$global:testroot = $PSScriptRoot
+$global:__pester_data = @{ }
+
+Remove-Module krbtgt -ErrorAction Ignore
+Import-Module "$PSScriptRoot\..\krbtgt.psd1"
+Import-Module "$PSScriptRoot\..\krbtgt.psm1" -Force
+
+# Need to import explicitly so we can use the configuration class
+Import-Module Pester
 
 Write-PSFMessage -Level Important -Message "Creating test result folder"
 $null = New-Item -Path "$PSScriptRoot\..\.." -Name TestResults -ItemType Directory -Force
@@ -31,6 +39,8 @@ $totalFailed = 0
 $totalRun = 0
 
 $testresults = @()
+$config = [PesterConfiguration]::Default
+$config.TestResult.Enabled = $true
 
 #region Run General Tests
 if ($TestGeneral)
@@ -38,21 +48,25 @@ if ($TestGeneral)
 	Write-PSFMessage -Level Important -Message "Modules imported, proceeding with general tests"
 	foreach ($file in (Get-ChildItem "$PSScriptRoot\general" | Where-Object Name -like "*.Tests.ps1"))
 	{
+		if ($file.Name -notlike $Include) { continue }
+		if ($file.Name -like $Exclude) { continue }
+		
 		Write-PSFMessage -Level Significant -Message "  Executing <c='em'>$($file.Name)</c>"
-		$TestOuputFile = Join-Path "$PSScriptRoot\..\..\TestResults" "TEST-$($file.BaseName).xml"
-    $results = Invoke-Pester -Script $file.FullName -Show $Show -PassThru -OutputFile $TestOuputFile -OutputFormat NUnitXml
+		$config.TestResult.OutputPath = Join-Path "$PSScriptRoot\..\..\TestResults" "TEST-$($file.BaseName).xml"
+		$config.Run.Path = $file.FullName
+		$config.Run.PassThru = $true
+		$config.Output.Verbosity = $Output
+		$results = Invoke-Pester -Configuration $config
 		foreach ($result in $results)
 		{
 			$totalRun += $result.TotalCount
 			$totalFailed += $result.FailedCount
-			$result.TestResult | Where-Object { -not $_.Passed } | ForEach-Object {
-				$name = $_.Name
+			$result.Tests | Where-Object Result -ne 'Passed' | ForEach-Object {
 				$testresults += [pscustomobject]@{
-					Describe = $_.Describe
-					Context  = $_.Context
-					Name	 = "It $name"
-					Result   = $_.Result
-					Message  = $_.FailureMessage
+					Block = $_.Block
+					Name  = "It $($_.Name)"
+					Result = $_.Result
+					Message = $_.ErrorRecord.DisplayErrorMessage
 				}
 			}
 		}
@@ -60,30 +74,33 @@ if ($TestGeneral)
 }
 #endregion Run General Tests
 
+$global:__pester_data.ScriptAnalyzer | Out-Host
+
 #region Test Commands
 if ($TestFunctions)
 {
-Write-PSFMessage -Level Important -Message "Proceeding with individual tests"
+	Write-PSFMessage -Level Important -Message "Proceeding with individual tests"
 	foreach ($file in (Get-ChildItem "$PSScriptRoot\functions" -Recurse -File | Where-Object Name -like "*Tests.ps1"))
 	{
 		if ($file.Name -notlike $Include) { continue }
 		if ($file.Name -like $Exclude) { continue }
 		
 		Write-PSFMessage -Level Significant -Message "  Executing $($file.Name)"
-		$TestOuputFile = Join-Path "$PSScriptRoot\..\..\TestResults" "TEST-$($file.BaseName).xml"
-    $results = Invoke-Pester -Script $file.FullName -Show $Show -PassThru -OutputFile $TestOuputFile -OutputFormat NUnitXml
+		$config.TestResult.OutputPath = Join-Path "$PSScriptRoot\..\..\TestResults" "TEST-$($file.BaseName).xml"
+		$config.Run.Path = $file.FullName
+		$config.Run.PassThru = $true
+		$config.Output.Verbosity = $Output
+		$results = Invoke-Pester -Configuration $config
 		foreach ($result in $results)
 		{
 			$totalRun += $result.TotalCount
 			$totalFailed += $result.FailedCount
-			$result.TestResult | Where-Object { -not $_.Passed } | ForEach-Object {
-				$name = $_.Name
+			$result.Tests | Where-Object Result -ne 'Passed' | ForEach-Object {
 				$testresults += [pscustomobject]@{
-					Describe = $_.Describe
-					Context  = $_.Context
-					Name	 = "It $name"
-					Result   = $_.Result
-					Message  = $_.FailureMessage
+					Block = $_.Block
+					Name  = "It $($_.Name)"
+					Result = $_.Result
+					Message = $_.ErrorRecord.DisplayErrorMessage
 				}
 			}
 		}