+++ Krbtgt/changelog.md
@@ -1,7 +1,19 @@
 ﻿# Changelog
+
+## 1.1.7 (2020-09-08)
+
+- Upd: All commands - New single item replication option: LDAP (instead of WinRM). LDAP new default
+- Upd: Get-KrbAccount - added "Credential" parameter
+- Upd: Reset-KrbPassword - added "Credential" parameter
+- Upd: Reset-KrbRODCPassword - added "Credential" parameter
+- Upd: Sync-KrbAccount - added "Credential" parameter
+- Upd: Test-KrbPasswordReset - added "Credential" parameter
+
 ## 1.0.1 (2019-07-15)
- - Fix: Test-KrbPasswordReset is not automatically picking up DCs if none were specified
- - Fix: Test-KrbPasswordReset ignores explicitly specified list of DCs
+
+- Fix: Test-KrbPasswordReset is not automatically picking up DCs if none were specified
+- Fix: Test-KrbPasswordReset ignores explicitly specified list of DCs
 
 ## 1.0.0 (2019-04-05)
- - New: Initial Release
+
+- New: Initial Release